Move some files in preparation for es6 modules
