+++ lib/buildenvsetup/base.js
