+++ lib/buildenvsetup/ceconan.js
@@ -24,7 +24,7 @@
 'use strict';
 
 const
-    BuildEnvSetupBase = require('./buildenvsetup-base'),
+    BuildEnvSetupBase = require('./base'),
     logger = require('../logger').logger,
     request = require('request'),
     fs = require('fs-extra'),
