+++ lib/buildenvsetup/cliconan.js
@@ -24,7 +24,7 @@
 'use strict';
 
 const
-    BuildEnvSetupBase = require('./buildenvsetup-base'),
+    BuildEnvSetupBase = require('./base'),
     exec = require('../exec'),
     logger = require('../logger').logger,
     fs = require('fs-extra'),
