+++ lib/storage/base.js
@@ -147,7 +147,7 @@ class StorageBase {
 }
 
 function storageFactory(storageSolution, compilerProps, awsProps, httpRootDir) {
-    const storage = require(`./storage-${storageSolution}`);
+    const storage = require(`./${storageSolution}`);
     return new storage(httpRootDir, compilerProps, awsProps);
 }
