+++ lib/storage/local.js
@@ -23,7 +23,7 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-const StorageBase = require('./storage').StorageBase,
+const StorageBase = require('./base').StorageBase,
     logger = require('../logger').logger,
     fs = require('fs-extra'),
     _ = require('underscore'),
