+++ lib/storage/null.js
@@ -23,59 +23,24 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-const logger = require('../logger').logger,
-    request = require('request'),
-    StorageBase = require('./storage').StorageBase,
-    util = require('util');
+const StorageBase = require('./base').StorageBase;
 
-class StorageRemote extends StorageBase {
-    constructor(httpRootDir, compilerProps) {
-        super(httpRootDir, compilerProps);
-
-        this.baseUrl = compilerProps.ceProps('remoteStorageServer');
-
-        const req = request.defaults({
-            baseUrl: this.baseUrl,
-        });
-
-        this.get = util.promisify(req.get);
-        this.post = util.promisify(req.post);
+class StorageNull extends StorageBase {
+    async storeItem(item) {
+        return item;
     }
 
-    async handler(req, res) {
-        let resp;
-        try {
-            resp = await this.post('/shortener', {
-                json: true,
-                body: req.body,
-            });
-        } catch (err) {
-            logger.error(err);
-            res.status(500);
-            res.end(err.message);
-            return;
-        }
-
-        const url = resp.body.url;
-        if (!url) {
-            res.status(resp.statusCode);
-            res.send(resp.body);
-            return;
-        }
-
-        const relativeUrl = url.substring(url.lastIndexOf('/z/') + 1);
-        const shortlink = `${req.protocol}://${req.get('host')}${this.httpRootDir}${relativeUrl}`;
-
-        res.send({url: shortlink});
-    }
-
-    async expandId(id) {
-        const resp = await this.get(`/api/shortlinkinfo/${id}`);
-
-        if (resp.statusCode !== 200) throw new Error(`ID ${id} not present in remote storage`);
-
+    async findUniqueSubhash() {
         return {
-            config: resp.body,
+            prefix: 'null',
+            uniqueSubHash: 'null',
+            alreadyPresent: true,
+        };
+    }
+
+    async expandId() {
+        return {
+            config: '{}',
             specialMetadata: null,
         };
     }
@@ -84,4 +49,4 @@ class StorageRemote extends StorageBase {
     }
 }
 
-module.exports = StorageRemote;
+module.exports = StorageNull;
