+++ lib/storage/remote.js
@@ -23,24 +23,59 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-const StorageBase = require('./storage').StorageBase;
+const logger = require('../logger').logger,
+    request = require('request'),
+    StorageBase = require('./base').StorageBase,
+    util = require('util');
 
-class StorageNull extends StorageBase {
-    async storeItem(item) {
-        return item;
+class StorageRemote extends StorageBase {
+    constructor(httpRootDir, compilerProps) {
+        super(httpRootDir, compilerProps);
+
+        this.baseUrl = compilerProps.ceProps('remoteStorageServer');
+
+        const req = request.defaults({
+            baseUrl: this.baseUrl,
+        });
+
+        this.get = util.promisify(req.get);
+        this.post = util.promisify(req.post);
     }
 
-    async findUniqueSubhash() {
-        return {
-            prefix: 'null',
-            uniqueSubHash: 'null',
-            alreadyPresent: true,
-        };
+    async handler(req, res) {
+        let resp;
+        try {
+            resp = await this.post('/shortener', {
+                json: true,
+                body: req.body,
+            });
+        } catch (err) {
+            logger.error(err);
+            res.status(500);
+            res.end(err.message);
+            return;
+        }
+
+        const url = resp.body.url;
+        if (!url) {
+            res.status(resp.statusCode);
+            res.send(resp.body);
+            return;
+        }
+
+        const relativeUrl = url.substring(url.lastIndexOf('/z/') + 1);
+        const shortlink = `${req.protocol}://${req.get('host')}${this.httpRootDir}${relativeUrl}`;
+
+        res.send({url: shortlink});
     }
 
-    async expandId() {
+    async expandId(id) {
+        const resp = await this.get(`/api/shortlinkinfo/${id}`);
+
+        if (resp.statusCode !== 200) throw new Error(`ID ${id} not present in remote storage`);
+
         return {
-            config: '{}',
+            config: resp.body,
             specialMetadata: null,
         };
     }
@@ -49,4 +84,4 @@ class StorageNull extends StorageBase {
     }
 }
 
-module.exports = StorageNull;
+module.exports = StorageRemote;
